+++ CHANGELOG.md
@@ -15,6 +15,7 @@ _This release is scheduled to be released on 2023-07-01._
 - Added tests for serveronly
 - Set Timezone `Europe/Berlin` in unit tests (needed for new formatTime tests)
 - Added no-param-reassign eslint rule and fix warnings
+- updatenotification: Added `sendUpdatesNotifications` feature. Broadcast update with `UPDATES` notification to other modules
 
 ### Removed